Fix minor issues with types within grid

+++ C++/data_types/grid.h
@@ -84,14 +84,14 @@ class Element
     }
 
     ElementEntity GetElement() const { return nodes_; };
-    std::uint8_t GetDimension() const { return dimension_; };
+    std::int8_t GetDimension() const { return dimension_; };
     FiniteElementOrder GetOrder() const { return order_; };
     const bool IsOnBoundary() const;
     void SetBoundaryBoolean(const bool is_on_boundary);
 
   private:
     ElementEntity nodes_{};
-    std::uint8_t dimension_{};
+    std::int8_t dimension_{};
     FiniteElementOrder order_{};
     bool is_on_boundary_{false};
 };
@@ -103,9 +103,9 @@ class Grid
     Grid(std::vector<Element> elements) : elements_(elements){};
 
   public:
-    std::uint8_t GetDimension() const { return dimension_; }
-    void SetDimension(const std::uint8_t& dimension) { dimension_ = dimension; }
-    std::uint64_t GetSize() const;
+    std::int8_t GetDimension() const { return dimension_; }
+    void SetDimension(const std::int8_t& dimension) { dimension_ = dimension; }
+    std::int32_t GetSize() const;
 
     std::vector<Element> GetElementsMutable() const { return elements_; }
     const std::vector<Element>& GetElements() const { return elements_; }
@@ -113,7 +113,7 @@ class Grid
 
   private:
     std::vector<Element> elements_{};
-    std::uint8_t dimension_{};
+    std::int8_t dimension_{};
 };
 
 }  // namespace geometry
